Correction of #111: -R (recursive) works only on level -1 subfolders
Recursive call to the directory + file list function is incorrect, as the path of sub-level is not taken into account. fd.cFilename contains only the name of the folder, but not the full path starting from the root. It needs to be added.

Currently, only level -1 from root level will be taken into account.